Yorumlara puanlama kısmı eklendi.Tarifi düzenleme ve silme kısmına kullanıcı kontrolü eklendi

+++ Yemek_Tarif/Controllers/RecipeController.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using Yemek_Tarif.Models;
 using Microsoft.EntityFrameworkCore;
+using Yemek_Tarif.Models;
 
 namespace Yemek_Tarif.Controllers
 {
@@ -26,22 +26,16 @@ namespace Yemek_Tarif.Controllers
         // GET: Recipe/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var recipe = await _context.Recipes
                 .Include(r => r.User)
                 .Include(r => r.Category)
-                .Include(r => r.Comments) // Yorumları da dahil ediyoruz
-                .ThenInclude(c => c.User) // Yorum yapan kullanıcıyı ekliyoruz
+                .Include(r => r.Comments)
+                .ThenInclude(c => c.User)
                 .FirstOrDefaultAsync(m => m.RecipeId == id);
 
-            if (recipe == null)
-            {
-                return NotFound();
-            }
+            if (recipe == null) return NotFound();
 
             return View(recipe);
         }
@@ -49,7 +43,6 @@ namespace Yemek_Tarif.Controllers
         // GET: Recipe/Create
         public IActionResult Create()
         {
-            // Kategorileri ViewBag üzerinden gönderiyoruz
             ViewBag.Categories = _context.Categories.ToList();
             return View();
         }
@@ -57,18 +50,18 @@ namespace Yemek_Tarif.Controllers
         // POST: Recipe/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("RecipeId,Title,Description,Ingredients,Instructions,UserId,CategoryId")] Recipe recipe)
+        public async Task<IActionResult> Create([Bind("RecipeId,Title,Description,Ingredients,Instructions,CategoryId")] Recipe recipe)
         {
             if (ModelState.IsValid)
             {
-                // DateCreated alanını ekliyoruz
+                recipe.UserId = (int)HttpContext.Session.GetInt32("UserId"); // Giriş yapan kullanıcı
                 recipe.DateCreated = DateTime.Now;
 
                 _context.Add(recipe);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            // Kategoriler tekrar gönderiliyor, çünkü model invalidse yeniden form render edilecek
+
             ViewBag.Categories = _context.Categories.ToList();
             return View(recipe);
         }
@@ -76,18 +69,19 @@ namespace Yemek_Tarif.Controllers
         // GET: Recipe/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var recipe = await _context.Recipes.FindAsync(id);
-            if (recipe == null)
+            if (recipe == null) return NotFound();
+
+            // Kullanıcı yetki kontrolü
+            int? userId = HttpContext.Session.GetInt32("UserId");
+            if (recipe.UserId != userId)
             {
-                return NotFound();
+                TempData["ErrorMessage"] = "Yetkiniz yok"; // Hata mesajını ekle
+                return RedirectToAction(nameof(Index));  // Kullanıcıyı ana sayfaya yönlendir
             }
 
-            // Kategorileri ViewBag üzerinden gönderiyoruz
             ViewBag.Categories = _context.Categories.ToList();
             return View(recipe);
         }
@@ -95,33 +89,35 @@ namespace Yemek_Tarif.Controllers
         // POST: Recipe/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("RecipeId,Title,Description,Ingredients,Instructions,DateCreated,UserId,CategoryId")] Recipe recipe)
+        public async Task<IActionResult> Edit(int id, [Bind("RecipeId,Title,Description,Ingredients,Instructions,DateCreated,CategoryId")] Recipe recipe)
         {
-            if (id != recipe.RecipeId)
+            if (id != recipe.RecipeId) return NotFound();
+
+            // Kullanıcı yetki kontrolü
+            int? userId = HttpContext.Session.GetInt32("UserId");
+            var originalRecipe = await _context.Recipes.AsNoTracking().FirstOrDefaultAsync(r => r.RecipeId == id);
+            if (originalRecipe == null || originalRecipe.UserId != userId)
             {
-                return NotFound();
+                TempData["ErrorMessage"] = "Yetkiniz yok"; // Hata mesajını ekle
+                return RedirectToAction(nameof(Index));  // Kullanıcıyı ana sayfaya yönlendir
             }
 
             if (ModelState.IsValid)
             {
                 try
                 {
+                    recipe.UserId = originalRecipe.UserId; // Kullanıcı ID'si değişmesin
                     _context.Update(recipe);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!RecipeExists(recipe.RecipeId))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!RecipeExists(recipe.RecipeId)) return NotFound();
+                    throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
+
             ViewBag.Categories = _context.Categories.ToList();
             return View(recipe);
         }
@@ -129,18 +125,21 @@ namespace Yemek_Tarif.Controllers
         // GET: Recipe/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var recipe = await _context.Recipes
                 .Include(r => r.User)
                 .Include(r => r.Category)
                 .FirstOrDefaultAsync(m => m.RecipeId == id);
-            if (recipe == null)
+
+            if (recipe == null) return NotFound();
+
+            // Kullanıcı yetki kontrolü
+            int? userId = HttpContext.Session.GetInt32("UserId");
+            if (recipe.UserId != userId)
             {
-                return NotFound();
+                TempData["ErrorMessage"] = "Yetkiniz yok"; // Hata mesajını ekle
+                return RedirectToAction(nameof(Index));  // Kullanıcıyı ana sayfaya yönlendir
             }
 
             return View(recipe);
@@ -152,6 +151,15 @@ namespace Yemek_Tarif.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var recipe = await _context.Recipes.FindAsync(id);
+
+            // Kullanıcı yetki kontrolü
+            int? userId = HttpContext.Session.GetInt32("UserId");
+            if (recipe == null || recipe.UserId != userId)
+            {
+                TempData["ErrorMessage"] = "Yetkiniz yok"; // Hata mesajını ekle
+                return RedirectToAction(nameof(Index));  // Kullanıcıyı ana sayfaya yönlendir
+            }
+
             _context.Recipes.Remove(recipe);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
@@ -164,16 +172,17 @@ namespace Yemek_Tarif.Controllers
         {
             if (string.IsNullOrEmpty(commentText))
             {
-                // Eğer yorum boş ise, kullanıcıya bir hata mesajı dönebiliriz
                 TempData["ErrorMessage"] = "Yorum alanı boş olamaz!";
                 return RedirectToAction("Details", new { id = recipeId });
             }
 
-            // Yorum eklemek için yeni bir Comment nesnesi oluşturuyoruz
+            int? userId = HttpContext.Session.GetInt32("UserId");
+            if (userId == null) return RedirectToAction("Login", "Account");
+
             var comment = new Comment
             {
                 RecipeId = recipeId,
-                UserId = (int)HttpContext.Session.GetInt32("UserId"), // Kullanıcı ID'sini oturumdan alıyoruz
+                UserId = (int)userId,
                 CommentText = commentText,
                 DateCreated = DateTime.Now
             };
@@ -181,7 +190,6 @@ namespace Yemek_Tarif.Controllers
             _context.Comments.Add(comment);
             await _context.SaveChangesAsync();
 
-            // Yorum eklendikten sonra, kullanıcının tarifin detay sayfasına yönlendirilmesi sağlanır
             return RedirectToAction("Details", "Recipe", new { id = recipeId });
         }
